+++ web/.prettierrc
@@ -1,5 +1,4 @@
 {
-  "endOfLine": "crlf",
   "semi": false,
   "singleQuote": true,
   "tabWidth": 2,
